+++ classes/Configuration/Configuration.php
@@ -14,11 +14,11 @@ class Configuration {
    * @param mixed $defaultValue
    * @return mixed
    */
-  public static function get($section, $key, $default_value = NULL) {
+  public static function get($section, $key, $defaultValue = NULL) {
     if (isset(self::$configuration[$section]) && isset(self::$configuration[$section][$key])) {
       return self::$configuration[$section][$key];
     } else {
-      return $default_value;
+      return $defaultValue;
     }
   }
 
@@ -61,7 +61,7 @@ class Configuration {
    * @param string $section
    * @param string $key
    */
-  public static function remove_key($section, $key) {
+  public static function remove($section, $key) {
     if (isset(self::$configuration[$section]) && isset(self::$configuration[$section][$key])) {
       unset(self::$configuration[$section][$key]);
     }
@@ -70,7 +70,7 @@ class Configuration {
   /**
    * @param string $section
    */
-  public static function remove_section($section) {
+  public static function removeSection($section) {
     if (isset(self::$configuration[$section])) {
       unset(self::$configuration[$section]);
     }
@@ -81,6 +81,23 @@ class Configuration {
    */
   public static function reset() {
     self::$configuration = [];
+  }
+
+  /**
+   * @param string $section
+   * @param string $key
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->remove() instead
+   */
+  public static function remove_key($section, $key) {
+    self::remove($section, $key);
+  }
+
+  /**
+   * @param string $section
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->removeSection() instead
+   */
+  public static function remove_section($section) {
+    self::removeSection($section);
   }
 
 }