+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 * Deprecated `AbstractModelCollection->remove_item()` because of its inconsistent naming. Use `AbstractModelCollection->remove()` instead.
 * Deprecated `BaseHttpAuthentication->check_authentication()`. Use `BaseHttpAuthentication->checkAuthentication()` instead.
 * Deprecated `BaseSessionInterface->clear_all()`. Use `BaseSessionInterface->clearAll()` instead.
+* Deprecated `Configuration::remove_key()`. Use `Configuration::remove()` instead.
+* Deprecated `Configuration::remove_section()`. Use `Configuration::removeSection()` instead.
 * Deprecated lower_camel_cased setters and getters in model classes. Use lowerCamelCased method names instead.
 
 ### Removed